photographyserver/sync_worker/pyproject.toml
[project]
name = "pgs-sync"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "pgs-db",
    "pillow>=11.2.1",
    "pydantic-settings>=2.9.1",
    "uvicorn[standard]>=0.34.3",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "httpx>=0.28.1",
    "psutil>=7.0.0",
    "pyright>=1.1.401",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-mock>=3.14.1",
    "ruff>=0.11.13",
    "sqlalchemy-stubs>=0.4",
]

[tool.uv.sources]
pgs-db = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = ["src/pgs_sync"]
[tool.hatch.build.targets.wheel]
packages = ["src/pgs_sync"]

[tool.pyright]
include = ["src/pgs_sync"]
exclude = ["tests"]
reportUnnecessaryTypeIgnoreComment = false
reportUnusedImport = false
